+++ logstash-filter-mongodb-guardium/src/main/java/com/ibm/guardium/connector/structures/Construct.java
@@ -1,4 +1,4 @@
-package com.ibm.guardium;
+package com.ibm.guardium.connector.structures;
 
 import java.util.ArrayList;