+++ logstash-filter-mongodb-guardium/src/main/java/com/ibm/guardium/connector/structures/Sentence.java
@@ -1,4 +1,4 @@
-package com.ibm.guardium;
+package com.ibm.guardium.connector.structures;
 
 import java.util.ArrayList;
 
@@ -8,7 +8,7 @@ public class Sentence {
     public ArrayList<Sentence> descendants = new ArrayList<>(); // TODO: implement?
     public ArrayList<String> fields = new ArrayList<>(); // TODO: implement? {key, value} objects
 
-    Sentence(String verb) {
+    public Sentence(String verb) {
         this.verb = verb;
     }
 }